Create LinkBuilder
